Refactor MC Quiz for JS and CSS
